% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasor.R
\name{phasor}
\alias{phasor}
\title{Function phasor can calculate a phasor}
\usage{
phasor(mod, degree)
}
\arguments{
\item{mod}{is a real number representing the module of a complex number.}

\item{degree}{is a degree number representing the argument of a complex number.}
}
\value{
a complex number
}
\description{
Function phasor can calculate a phasor
}
\examples{
phasor(1,30)
}
