% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasor.R
\name{rephasor}
\alias{rephasor}
\title{Function rephasor can develop a phasor}
\usage{
rephasor(complex)
}
\arguments{
\item{complex}{is a complex number to be calculated.}
}
\value{
a atomic numeric vector representing a phasor
}
\description{
Function rephasor can develop a phasor
}
\examples{
rephasor(sqrt(3)/2+0.5i)
}
